Added: Implemented a method to calculate the distance of a line in the MapCanvas. (Currently calculates the horizontal line from window start to end.)
Added: Multithreading for loading of OSM files. Decoupled Loader and Creator in order to do so.

Added: New class ProgressInputStream and Interface InputStreamListener to be able to monitor the current progress of the XMLStreamReader.

GUI: Added a ProgressIndicator while loading an OSM-file.

Changed: Extension of themes changed from .theme to .mtheme to avoid clash with Windows Theme Files.