Bug Fixed: Changing theme after two map loads caused default and dark mode to be mixed.
Bug Fixed: A new change listener was added to zoomSlider at each map load leading to serious problems when zooming in.

Changed: Changed the way the Controller saves the list of themes which can be used. Each menu item now has a path to its theme, instead of using a HashMap.